+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Adding `shlex_split`, `cmd` and `shell` moonblade functions.
 * Adding `aarch64-apple-darwin` and `aarch64-unknown-linux-gnu` to CI builds.
 * Adding `to_fixed` moonblade function.
+* Adding decimal places optional argument to `ratio` & `percentage` aggregation functions.
 
 *Fixes*